Cleanup misc Design classes
