+++ src/System.Windows.Forms.Design/src/System/Windows/Forms/Design/AxImporter.cs
@@ -10,10 +10,7 @@ using System.Runtime.InteropServices.ComTypes;
 namespace System.Windows.Forms.Design
 {
     /// <summary>
-    ///     <para>
-    ///         Imports ActiveX controls and generates a wrapper that can be accessed by a
-    ///         designer.
-    ///     </para>
+    /// Imports ActiveX controls and generates a wrapper that can be accessed by a designer.
     /// </summary>
     public class AxImporter
     {
@@ -21,60 +18,91 @@ namespace System.Windows.Forms.Design
         {
             throw new NotImplementedException(SR.NotImplementedByDesign);
         }
+
         public sealed class Options
         {
+            /// <summary>
             /// The flag that controls when we sign the generated assemblies.
+            /// </summary>
             public bool delaySign = false;
 
+            /// <summary>
             /// Flag that controls whether we are to generate sources for the ActiveX control wrapper.
+            /// </summary>
             public bool genSources = false;
 
+            /// <summary>
             /// Regardless of which version of the library is registered, if at all, use the 
             /// path supplied on the command line for AxImp to generate the Windows Forms wrappers.
-            [SuppressMessage("Microsoft.Design", "CA1051:DoNotDeclareVisibleInstanceFields")]
-            //This class is used to communicate option values between components, it does not have any functionality which might modify these values.
+            /// </summary>
+            [SuppressMessage("Microsoft.Design", "CA1051:DoNotDeclareVisibleInstanceFields", Justification = "This class is used to communicate option values between components, it does not have any functionality which might modify these values.")]
             public bool ignoreRegisteredOcx;
 
+            /// <summary>
             /// The file containing the strong name key container for the generated assemblies.
+            /// </summary>
             public string keyContainer = null;
 
+            /// <summary>
             /// The file containing the strong name key for the generated assemblies.
+            /// </summary>
             public string keyFile = null;
 
+            /// <summary>
             /// The strong name used for the generated assemblies.
+            /// </summary>
             public StrongNameKeyPair keyPair = null;
 
+            /// <summary>
             /// Flag that controls whether we should output errors in the MSBuild format.
+            /// </summary>
             public bool msBuildErrors = false;
 
+            /// <summary>
             /// Flag that controls whether we should show the logo.
+            /// </summary>
             public bool noLogo = false;
 
+            /// <summary>
             /// The output directory for all the generated assemblies.
+            /// <summary>
             public string outputDirectory = null;
 
+            /// <summary>
             /// The path-included filename of type library containing the definition of the ActiveX control.
+            /// </summary>
             public string outputName = null;
 
+            /// <summary>
             /// The flag that controls whether we try to overwrite existing generated assemblies.
+            /// </summary>
             public bool overwriteRCW = false;
+
+            /// <summary>
             /// The public key used to sign the generated assemblies.
+            /// </summary>
             public byte[] publicKey = null;
 
+            /// <summary>
             /// The object that allows us to resolve types and references needed to generate assemblies.
+            /// </summary>
             public IReferenceResolver references = null;
 
+            /// <summary>
             /// Flag that controls the output generated.
+            /// </summary>
             public bool silentMode = false;
 
+            /// <summary>
             /// Flag that controls the output generated.
+            /// </summary>
             public bool verboseMode = false;
         }
 
         /// <summary>
-        ///     The Reference Resolver service will try to look through the references it can obtain,
-        ///     for a reference that matches the given criterion. For now, the only kind of references
-        ///     it can look for are COM (RCW) references and ActiveX wrapper references.
+        /// The Reference Resolver service will try to look through the references it can obtain,
+        /// for a reference that matches the given criterion. For now, the only kind of references
+        /// it can look for are COM (RCW) references and ActiveX wrapper references.
         /// </summary>
         public interface IReferenceResolver
         {
@@ -87,37 +115,4 @@ namespace System.Windows.Forms.Design
             string ResolveActiveXReference(UCOMITypeLib typeLib);
         }
     }
-}
-
-namespace System.Runtime.InteropServices
-{
-
-    // UCOMITypeLib is not yet ported to interop on core.
-
-    [Guid("00020402-0000-0000-C000-000000000046")]
-    [InterfaceType(ComInterfaceType.InterfaceIsIUnknown)]
-    [ComImport]
-    public interface UCOMITypeLib
-    {
-        [PreserveSig]
-        int GetTypeInfoCount();
-
-        //void GetTypeInfo(int index, out UCOMITypeInfo ppTI);
-        void GetTypeInfoType(int index, out TYPEKIND pTKind);
-
-        //void GetTypeInfoOfGuid(ref Guid guid, out UCOMITypeInfo ppTInfo);
-        void GetLibAttr(out IntPtr ppTLibAttr);
-
-        //void GetTypeComp(out UCOMITypeComp ppTComp);
-        void GetDocumentation(int index, out string strName, out string strDocString, out int dwHelpContext,
-            out string strHelpFile);
-
-        [return: MarshalAs(UnmanagedType.Bool)]
-        bool IsName([MarshalAs(UnmanagedType.LPWStr)] string szNameBuf, int lHashVal);
-
-        //void FindName([MarshalAs(UnmanagedType.LPWStr)] String szNameBuf, int lHashVal, [MarshalAs(UnmanagedType.LPArray), Out] UCOMITypeInfo[] ppTInfo, [MarshalAs(UnmanagedType.LPArray), Out] int[] rgMemId, ref Int16 pcFound);
-        [PreserveSig]
-        void ReleaseTLibAttr(IntPtr pTLibAttr);
-    }
-
 }
