+++ src/System.Windows.Forms.Design/src/System/Windows/Forms/Design/StandardGroups.cs
@@ -2,12 +2,9 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Diagnostics.CodeAnalysis;
-
 namespace System.Windows.Forms.Design
 {
-    [SuppressMessage("Microsoft.Performance", "CA1812:AvoidUninstantiatedInternalClasses")]
-    internal class StandardGroups
+    internal static class StandardGroups
     {
         public const string Code = "Code";
         public const string ZORder = "ZOrder";
