+++ src/System.Windows.Forms.Design/src/System/Windows/Forms/Design/ThemedScrollbarMode.cs
@@ -2,8 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Collections;
-
 namespace System.Windows.Forms.Design
 {
     public enum ThemedScrollbarMode
@@ -23,10 +21,4 @@ namespace System.Windows.Forms.Design
         /// </summary>
         OnlyTopLevel = 3
     };
-
-    public struct ThemedScrollbarWindow
-    {
-        public IntPtr Handle;
-        public ThemedScrollbarMode Mode;
-    }
 }
